requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
youtube-transcript-api
langchain
langchain-openai
langchain-community
openai
mangum
faiss-cpu
langchain_pinecone
pinecone

